Add test for caching.
This requires changing the way stubs work, so that they only simulate
actual HTTP calls, not cache hits.
